1st

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { Task } from "./task.entity";
 import { TaskStatus } from "./task-status.enum";
 import { GetTasksFilterDto } from "./dto/get-tasks-filter.dto";
+import { User } from "../auth/user.entity";
 
 @Injectable()
 export class TasksService {
@@ -14,24 +15,24 @@ export class TasksService {
     ) {
     }
 
-    getTasks(filterDto :GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    getTasks(filterDto :GetTasksFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({id, userId: user.id});
         if (!found) {
             throw new NotFoundException(`task id=${id} not found`);
         }
         return found;
     }
 
-    async createTask(dto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(dto);
+    async createTask(dto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(dto, user);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         if (!task) {
             throw new NotFoundException(`task id=${id} is not found`);
         }
@@ -40,8 +41,8 @@ export class TasksService {
     }
 
 
-    async deleteTaskById(id: number): Promise<boolean> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<boolean> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
         if(result.affected === 0) {
             throw new NotFoundException(`Task id=${id} not found.`);
         }
